Change player Id to use boxed Long type to allow nullable values inside BombState.

+++ src/test/java/uk/oczadly/karl/csgsi/state/BombStateTest.java
@@ -24,7 +24,7 @@ public class BombStateTest extends GameStateBaseTest {
         
         assertEquals(BombState.BombStatus.CARRIED, state.getPhase().get());
         assertEquals(new Coordinate(3084.00, 127.00, 1613.03), state.getPosition());
-        assertEquals(76561197960265734L, state.getPlayerId());
+        assertEquals(76561197960265734L, (long)state.getPlayerId());
     }
